Make VS Code and Ant play nicely together.
The "Language Support for Java™ by Red Hat" compiles the code while
you are typing to help provide IntelliSense.  The downside of this, it
will produce .class files even if there are fatal compile errors.

The Ant build system will see that the .class file is newer than the
.java file and will skip rebuilding it.

One way to solve it is to use separate outputPath directories for And
and VS Code.  I've kept the same path for VS Code and told Ant to
place the output in "build-ant".